인스펙터 변동

+++ Assets/Member/NSY/Script/TakeScreenShot/Screenshot.cs
@@ -302,6 +302,7 @@ public class Screenshot : MonoBehaviour
             //pc 저장
             byte[] byteArray = screenshotTexture1.EncodeToPNG(); // 이미지 저장
             byte[] byteArray2 = screenshotTexture2.EncodeToPNG(); // 이미지 저장
+
             byte[] byteArray3 = screenshotTexture3.EncodeToPNG(); // 전체 이미지 저장
                                                                   //
                                                                   //Ver 2  . 무지성 저장 카메라 영역에대한
@@ -321,6 +322,7 @@ public class Screenshot : MonoBehaviour
                        
                         Material1.SetTexture("_MainTex", Texutre2);
                         Material2.SetTexture("_MainTex", Texutre1);
+
                         Material3.SetTexture("_MainTex", Texutre3);
                     }
                     else
@@ -328,6 +330,7 @@ public class Screenshot : MonoBehaviour
                       
                         Material1.SetTexture("_MainTex", Texutre2);
                         Material2.SetTexture("_MainTex", Texutre1);
+
                         Material3.SetTexture("_MainTex", Texutre3);
                     }
                 }
@@ -347,7 +350,7 @@ public class Screenshot : MonoBehaviour
                 System.IO.File.WriteAllBytes(TotalPath, screenshotTexture1.EncodeToPNG());
                 System.IO.File.WriteAllBytes(TotalPath, screenshotTexture2.EncodeToPNG());
             }
-            else
+            else if(Directory.Exists(FolderPath) == true)
             {
                 System.IO.File.WriteAllBytes(TotalPath, screenshotTexture1.EncodeToPNG());
                 System.IO.File.WriteAllBytes(TotalPath, screenshotTexture2.EncodeToPNG());
@@ -360,7 +363,7 @@ public class Screenshot : MonoBehaviour
                 Directory.CreateDirectory(publicFolder);
                 System.IO.File.WriteAllBytes(publicTotalPath, screenshotTexture3.EncodeToPNG()); //publicTotalPath
             }
-            else
+            else if (Directory.Exists(publicFolder) == true)
             {
                 System.IO.File.WriteAllBytes(publicTotalPath, screenshotTexture3.EncodeToPNG()); //publicTotalPath
             }
@@ -371,6 +374,15 @@ public class Screenshot : MonoBehaviour
 
 
 
+    }
+
+    private void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.M))
+        {
+            Debug.Log("임소정씨발련");
+            OnSceenShotEvent();
+        }
     }
 
     // Update is called once per frame
@@ -384,7 +396,10 @@ public class Screenshot : MonoBehaviour
 
         }
         OnPostRender();
-        _WillFakeScreenShot = false;
+         _WillFakeScreenShot = false;
+
+        
+
     }
